requirements.txt
dotenv == 0.9.9
discord.py == 2.6.4
pyyaml == 6.0.1
googletrans == 4.0.2